Cleanup OpenAI stream test.
